# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  version: 1.0.0
  description: API documentation for the E-commerce platform
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: string
          format: uuid
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponse:
      type: object
      properties:
        type:
          type: string
          enum: [OK, ERROR]
        message:
          type: string
        data:
          type: object
        meta:
          type: object
    PaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                total:
                  type: integer
                page:
                  type: integer
                limit:
                  type: integer
                totalPages:
                  type: integer
                hasNextPage:
                  type: boolean
                hasPrevPage:
                  type: boolean
            filters:
              type: object
              properties:
                applied:
                  type: object
                available:
                  type: object
            sort:
              type: object
              properties:
                by:
                  type: string
                order:
                  type: string
                  enum: [asc, desc]
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get products or a specific product by ID
      parameters:
        - in: query
          name: id
          schema:
            type: string
            format: uuid
          description: ID of a specific product to fetch
        - in: query
          name: product_ids
          schema:
            type: string
          description: Array of product IDs (JSON string or comma-separated) to fetch specific products
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: inStock
          schema:
            type: boolean
        - in: query
          name: sortBy
          schema:
            type: string
            default: createdAt
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Product data
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/PaginatedResponse'
              examples:
                singleItem:
                  value:
                    type: "OK"
                    data:
                      product:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        name: "Smartphone X"
                        description: "Latest smartphone with advanced features"
                        price: 999.99
                        categoryId: "123e4567-e89b-12d3-a456-426614174001"
                        images: ["https://example.com/phone1.jpg"]
                        inStock: true
                multipleItems:
                  value:
                    type: "OK"
                    data:
                      products: []
                    pagination:
                      total: 50
                      page: 1
                      limit: 10
                      totalPages: 5
                      hasNextPage: true
                      hasPrevPage: false
                    filters:
                      applied:
                        categoryId: "123e4567-e89b-12d3-a456-426614174001"
                        minPrice: 500
                        maxPrice: 1000
                      available:
                        categoryId: "123e4567-e89b-12d3-a456-426614174001"
                        minPrice: 500
                        maxPrice: 1000
                        inStock: true
                        product_ids: "123e4567-e89b-12d3-a456-426614174000,123e4567-e89b-12d3-a456-426614174001"
                    sort:
                      by: "price"
                      order: "asc"
                productIdsResponse:
                  value:
                    type: "OK"
                    data:
                      products: []
                      requestedIds: ["123e4567-e89b-12d3-a456-426614174000", "123e4567-e89b-12d3-a456-426614174001"]
                      foundIds: ["123e4567-e89b-12d3-a456-426614174000"]
                      missingIds: ["123e4567-e89b-12d3-a456-426614174001"]

    post:
      tags:
        - Products
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - description
                - price
                - sku
                - categoryId
                - files
              properties:
                name:
                  type: string
                  description: Name of the product
                description:
                  type: string
                  description: Description of the product
                price:
                  type: number
                  description: Price of the product
                sku:
                  type: string
                  description: Stock Keeping Unit (SKU) of the product
                categoryId:
                  type: string
                  format: uuid
                  description: ID of the category the product belongs to
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Image files to upload (at least one required)
                brandId:
                  type: string
                  format: uuid
                  description: ID of the brand (optional)
                compareAtPrice:
                  type: number
                  description: Compare at price (optional)
                costPrice:
                  type: number
                  description: Cost price (optional)
                stock:
                  type: number
                  description: Stock quantity (optional, default 0)
                lowStockThreshold:
                  type: number
                  description: Low stock threshold (optional)
                availableSizes:
                  type: array
                  items:
                    type: string
                  description: Available sizes (optional)
                colors:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      code:
                        type: string
                  description: Colors (optional)
                weight:
                  type: object
                  properties:
                    value:
                      type: number
                    unit:
                      type: string
                      enum: [g, kg, lb, oz]
                  description: Weight (optional)
                dimensions:
                  type: object
                  properties:
                    length:
                      type: number
                    width:
                      type: number
                    height:
                      type: number
                    unit:
                      type: string
                      enum: [cm, m, in]
                  description: Dimensions (optional)
                tags:
                  type: array
                  items:
                    type: string
                  description: Tags (optional)
                status:
                  type: string
                  enum: [draft, active, archived, out_of_stock]
                  description: Product status (optional, default draft)
                isFeatured:
                  type: boolean
                  description: Whether the product is featured (optional, default false)
                seo:
                  type: object
                  properties:
                    title:
                      type: string
                    description:
                      type: string
                    keywords:
                      type: array
                      items:
                        type: string
                  description: SEO details (optional)
                ratings:
                  type: object
                  properties:
                    average:
                      type: number
                    count:
                      type: number
                  description: Ratings (optional)
                barcode:
                  type: string
                  description: Barcode (optional)
                media:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                      type:
                        type: string
                        enum: [image, video]
                      alt:
                        type: string
                  description: Additional media (optional)
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/Product'

    put:
      tags:
        - Products
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Products
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to delete
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...